Jenkinsfile: bump memory and CPU for K8s 1.12 tests
The Istio tests put extreme load on the VMs:

```
 21:11:27 up 20 min,  0 users,  load average: 17.89, 15.56, 8.53
```

This is because the test runs a ton of pods, each with their own Envoy sidecar
proxy, which uses considerable resources. Bump the resources in the VMs
accordingly to reduce the amount of load on the system.

Signed-off by: Ian Vernon <ian@cilium.io>