test: add new skipLogs option to CI
Before, log gathering was skipped if `-cilium.holdEnvironment=true`. Instead,
decouple log gathering from holding the environment for local debugging so that
if developers desire, they can still gather logs while keeping the local
environment preserved. This is useful in the case where the output from
`cilium monitor` wants to be introspected, as it can reveal transient behavior
in the datapath that is incorrect which is impossible to discover otherwise.

Signed-off by: Ian Vernon <ian@cilium.io>